move canvas payload function to a shared file
